arc4.js to arc4.ts conversion
